requirements.txt
asgiref==3.11.1
Django==6.0.3
django-environ==0.13.0
djangorestframework==3.16.1
numpy==2.4.2
pgvector==0.4.2
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2025.3
PyMuPDF
google-generativeai
python-telegram-bot
celery
redis
gunicorn
whitenoise
